refactor: Repository 추가함에 따라 리팩토링 진행
