+++ src/main/java/subway/controller/LineController.java
@@ -1,10 +1,10 @@
-package subway.ui;
+package subway.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import subway.application.LineService;
-import subway.application.SubwayMapService;
-import subway.domain.entity.LineEntity;
+import subway.service.LineService;
+import subway.service.SubwayMapService;
+import subway.entity.LineEntity;
 import subway.dto.line.LineRequest;
 import subway.dto.station.LineMapResponse;
 
@@ -14,6 +14,16 @@ import java.util.List;
 @RestController
 @RequestMapping("/lines")
 public class LineController {
+
+    // TODO: 테스트 추가, 도메인 메서드 및 변수명 변경, README 수정, 메서드 분리, 예외 핸들링, @Valid 적용
+
+    /**
+     * 1. 리팩토링 (코드 분리 + 도메인 메서드 명 변경)
+     * 2. 테스트
+     * 3. 예외 핸들링
+     * 4. @Valid 적용
+     * 5. README 수정
+     */
 
     private final LineService lineService;
     private final SubwayMapService subwayMapService;
